+++ src/physics/Physics.hpp
@@ -8,8 +8,9 @@
 
 #ifndef PHYSICS_HPP
 #define PHYSICS_HPP
-#include <glm/glm.hpp>
 
+#include <glm/glm.hpp>
+#include "Collider.hpp"
 
 /**
  * @class Physics
@@ -108,73 +109,6 @@ public:
    * @author Landry GIGANT
    */
   static void applyImpulse(Vector3 &vel, const Vector3 &impulse);
- };
-
- class ICollider {
- public:
-  virtual ~ICollider() = default;
-
-  virtual bool Collide(ICollider *collider) = 0;
- };
-
- /**
- * @class Box
- * @brief A class that represents a box in 3D space.
- * @details This class contains the position and size of the box.
- * @version v0.1.0
- * @since v0.1.0
- * @author Landry GIGANT
- */
- class Box : public ICollider {
- public:
-  /**
-   * @brief The position of the box in 3D space.
-   */
-  Vector3 position;
-
-  /**
-   * @brief The size of the box in 3D space.
-   */
-  Vector3 size;
-
-  /**
-   * @brief Constructor for the Box class.
-   * @param position The position of the box in 3D space.
-   * @param size The size of the box in 3D space.
-   */
-  Box(Vector3 position, Vector3 size) : position(position), size(size) {};
-
-  bool Collide(ICollider *collider) override;
- };
-
-/**
- * @class Sphere
- * @brief A class that represents a sphere in 3D space.
- * @details This class contains the position and radius of the sphere.
- * @version v0.1.0
- * @since v0.1.0
- * @author Landry GIGANT
- */
- class Sphere : public ICollider {
- public:
-  /**
-   * @brief The position of the sphere in 3D space.
-   */
-  Vector3 position;
-
-  /**
-   * @brief The radius of the sphere.
-   */
-  float radius;
-
-  /**
-   * @brief Constructor for the Sphere class.
-   * @param position The position of the sphere in 3D space.
-   * @param radius The radius of the sphere.
-   */
-  Sphere(Vector3 position, float radius) : position(position), radius(radius) {};
-
-  bool Collide(ICollider *collider) override;
  };
 
 /**
